utils: move funcs - ensure file exists, rewrite get_file_sizes func

+++ util/utils.py
@@ -73,15 +73,17 @@ def get_videos(source_dir):
     return video_files
 
 
+# move small files to small_files directory using get_videos function
 def move_small(source_dir, size: int):
     """
-    This function takes a source directory and a size in megabytes as arguments and moves all files in the source directory
-    that are less than the size to a destination directory named small_files
+    It takes a source directory and a size in megabytes as arguments, and moves all files smaller than the size to a new
+    directory called small_files
 
-    :param source_dir: the directory where the files are currently located
-    :param size: the size in megabytes that you want to move
+    :param source_dir: the directory where the files are located
+    :param size: the size of the file in megabytes
     :type size: int
     """
+
     # get current working directory and create dest_dir named small_files
     cwd = os.getcwd()
     dest_dir = os.path.join(cwd, 'small_files')
@@ -91,27 +93,27 @@ def move_small(source_dir, size: int):
     files = get_videos(source_dir)
     # loop through the files
     for file in files:
-        # get the file size
-        file_size = os.path.getsize(os.path.join(source_dir, file))
-        # if the file size is greater than 100 megabytes
-
-        # convert the bytes to megabytes
-        file_size_mb = file_size / 1000000
-
-        if file_size_mb < size:
-            # create a directory if it doesn't exist
-            if not os.path.exists(dest_dir):
-                os.makedirs(dest_dir)
-            # move the file to the destination directory
-            shutil.move(os.path.join(source_dir, file), dest_dir)  # move the file to the destination directory
+        # check if file exists
+        if os.path.exists(os.path.join(source_dir, file)):
+            # get the file size
+            file_size = os.path.getsize(os.path.join(source_dir, file))
+            # convert the file size to megabytes
+            file_size_mb = file_size / 1000000
+            # if the file size is smaller than 100 megabytes
+            if file_size_mb < size:
+                # create a directory if it doesn't exist
+                if not os.path.exists(dest_dir):
+                    os.makedirs(dest_dir)
+                # move the file to the destination directory
+                shutil.move(os.path.join(source_dir, file), dest_dir)
 
 
 def move_large(source_dir, size: int):
     """
-    This function takes a source directory and a size in megabytes as arguments, and moves all files in the source directory
-    that are larger than the specified size to a destination directory named large_files.
+    It takes a source directory and a size in megabytes as arguments, and moves all files larger than the specified size to
+    a new directory called large_files
 
-    :param source_dir: the directory where the files are currently located
+    :param source_dir: the directory where the files are located
     :param size: the size of the file in megabytes
     :type size: int
     """
@@ -124,40 +126,36 @@ def move_large(source_dir, size: int):
     files = get_videos(source_dir)
     # loop through the files
     for file in files:
-        # get the file size
-        file_size = os.path.getsize(os.path.join(source_dir, file))
-        # if the file size is greater than 100 megabytes
-        # convert the file size to megabytes
-        file_size_mb = file_size / 1000000
+        # check if file exists
+        if os.path.exists(os.path.join(source_dir, file)):
+            # get the file size
+            file_size = os.path.getsize(os.path.join(source_dir, file))
+            # convert the file size to megabytes
+            file_size_mb = file_size / 1000000
+            # if the file size is smaller than 100 megabytes
+            if file_size_mb > size:
+                # create a directory if it doesn't exist
+                if not os.path.exists(dest_dir):
+                    os.makedirs(dest_dir)
+                # move the file to the destination directory
+                shutil.move(os.path.join(source_dir, file), dest_dir)
 
-        if file_size_mb > size:
-            # create a directory if it doesn't exist
-            if not os.path.exists(dest_dir):
-                os.makedirs(dest_dir)
-            # move the file to the destination directory
-            shutil.move(os.path.join(source_dir, file), dest_dir)  # move the file to the destination directory
 
-
-# get file name and file sizes in megabytes using get_files function
 def get_file_sizes(source_dir):
     """
-    It takes a directory as an argument, gets the files in that directory, creates a list to store the file sizes, loops
-    through the files, gets the file size, converts the bytes to megabytes, and adds the file name and file size to the list
+    It takes a directory as input and returns a list of files in that directory along with their sizes.
 
-    :param source_dir: the directory where the files are located
-    :return: A list of tuples containing the file name and file size in megabytes.
+    :param source_dir: The directory from which you want to get the file sizes
     """
-    # get the files in the source directory
-    files = get_videos(source_dir)
-    # create a list to store the file sizes
-    file_sizes = []
-    # loop through the files
-    for file in files:
-        # get the file size
-        file_size = os.path.getsize(os.path.join(source_dir, file))
-        # convert the bytes to megabytes
-        file_size_mb = file_size / 1000000
-        # add the file name and file size to the list
-        file_sizes.append((file, file_size_mb))
-    # return the list
-    return file_sizes
+    f = lambda x: os.path.isfile(os.path.join(source_dir, x))
+    files_list = filter(f, os.listdir(source_dir))
+
+    # Create a list of files in directory along with the size
+    size_of_file = [
+        (f, os.stat(os.path.join(source_dir, f)).st_size)
+        for f in files_list
+    ]
+    # Iterate over list of files along with size
+    # and print them one by one.
+    for f, s in size_of_file:
+        print("{}: {} MB".format(f, round(s / (1024 * 1024), 3)))
